Modified AnsibleRunner to handle having a base directory.
If other code requests a temporary directory as a base (currently
AnsibleResourceModelSource does this), the base directory will be marked
for deletion unless requested to keep directory. If no base directory is
specified then a temporary directory will be used and marked for
deletion.

The default behaviour matches the current temporary directory behaviour
for a working directory. This will however allow a user to specify a
working directory.